Project ipynb File
Task 1: Clustering on Synthetic Data
In this task, the K-means and hierarchical clustering methods are applied to eight different datasets. The performance of clustering algorithms is evaluated using external validation metrics. The repository includes code to plot the data points in 2D or 3D, colored according to the original class and class allocated by the clustering algorithm.

Task 2: Clustering on World Indicators Dataset
In this task, the K-means and hierarchical clustering methods are applied to group similar countries together based on selected attributes. Internal validation metrics are used to report the cluster quality, and the best clustering solution is reported. The repository includes three different scatter plots of the world indicators dataset, colored according to the group.

This repository is helpful for anyone who wants to learn and understand the implementation of clustering algorithms on synthetic and real-world data, and evaluate their performance using external and internal validation metrics.